requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
frozendict==2.4.5
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
multitasking==0.0.11
mysql-connector-python==9.0.0
numpy==1.24.4
openai==1.51.0
pandas==2.0.3
peewee==3.17.6
platformdirs==4.3.6
psycopg2-binary==2.9.1
pydantic==2.9.2
pydantic-core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
werkzeug==3.0.4
yarl==1.13.1
yfinance==0.2.44
zipp==3.20.2
